feat: move loading component to item/id page
